requirements.txt
torch
torchvision
timm
avalanche-lib
pandas
matplotlib
seaborn
PyPDF2
